+++ test/bmi_test_utils.h
@@ -7,8 +7,8 @@
 /* Define some expected constants */
 #define EXPECTED_COMPONENT_NAME "CFE - Conceptual Functional Equivalent"
 #define EXPECTED_INPUT_VAR_COUNT 2
-#define EXPECTED_OUTPUT_VAR_COUNT 20
-#define EXPECTED_TOTAL_VAR_COUNT 22  // This should equal the sum of the input and output counts
+#define EXPECTED_OUTPUT_VAR_COUNT 23
+#define EXPECTED_TOTAL_VAR_COUNT 25
 #define EXPECTED_MODULE_START_TIME 0.0  // Assume the standard here
 #define EXPECTED_TIME_STEP_SIZE 3600.0   // Assume expected time step is 1 hour, in seconds